requirements.txt
# === CORE ===
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.4
pydantic_core==2.33.2
typing-extensions==4.13.2
azure-storage-blob==12.20.0

# === ASGI / AIO ===
anyio==4.9.0
sniffio==1.3.1
starlette==0.46.2
aiohttp==3.11.18
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
frozenlist==1.6.0
multidict==6.4.3
yarl==1.20.0

# === JWT + SECURITY ===
PyJWT==2.10.1
python-jose==3.5.0
bcrypt==4.3.0
passlib==1.7.4
passlib[bcrypt]==1.7.4
cryptography==44.0.3
cffi==1.17.1
pycparser==2.22
itsdangerous==2.2.0

# === DATABASES ===
mysql-connector-python==9.3.0
pyodbc==5.2.0
SQLAlchemy==2.0.36
alembic==1.14.0

# === GOOGLE / FIREBASE ===
firebase-admin==6.8.0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-firestore==2.20.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-api-core==2.24.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
protobuf==5.29.4
proto-plus==1.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
httplib2==0.22.0
uritemplate==4.1.1
rsa==4.9.1
cachetools==5.5.2
blinker==1.9.0
requests-toolbelt==0.10.1

# === WEB3 ===
web3==7.11.0
eth-abi==5.2.0
eth-account==0.13.7
eth-rlp==2.2.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.0
hexbytes==1.3.0
rlp==4.1.0
toolz==1.0.0
cytoolz==1.0.1
parsimonious==0.10.0

# === CRYPTO ===
ecdsa==0.19.1
pycryptodome==3.22.0
jwcrypto==1.5.6

# === UTILITIES ===
requests==2.32.3
urllib3>=1.26.20,<2
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
regex==2024.11.6
pyparsing==3.2.3
MarkupSafe==3.0.2
click==8.1.8
colorama==0.4.6
pymssql==2.3.6
python-multipart==0.0.9